backend/requirements.txt
django
djangorestframework
django-rest-auth
django-allauth
django-extensions
Werkzeug
pillow
celery
-e hg+https://vojtek@bitbucket.org/iamFIREcracker/osaic#egg=osaic
django-cors-headers
easy-thumbnails